读文件写入数据库 路径和数据库改为配置文件内容 密码验证

+++ db/db_manage.py
@@ -1,6 +1,4 @@
 import base64
-import secrets
-
 import pymysql
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import hashes, padding
@@ -15,6 +13,7 @@ class DB():
         self.key=filepath
         self.config=Config()
         self.config_ini=self.config.read_config()
+        self.database_name = self.config_ini['db_set']['database_name']
         self.conn = pymysql.connect(
             host="localhost",
             port=int(self.config_ini['db_set']['port']),
@@ -59,7 +58,7 @@ class DB():
         return encrypted_data
 
     def table_to_file(self):
-        table_name='func'
+        table_name=self.config_ini['db_set']['form_2']
         ff=self.config_ini['main_project']['project_path']+self.config_ini['db']['back_up']
         with open(ff, 'w') as file:
             self.cursor.execute(f'SELECT * FROM {table_name}')
@@ -73,19 +72,30 @@ class DB():
                 file.write(name + "\t" + level + "\t" + solution + '\n')
 
     def table_clear(self,table_name):
-        table_name = table_name
-        truncate_query = f"TRUNCATE TABLE {table_name};"
-        self.cursor.execute(truncate_query)
-        self.conn.commit()
+        query = f"SELECT TABLE_NAME FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_SCHEMA = '{self.database_name}' AND TABLE_NAME = '{table_name}'"
+        self.cursor.execute(query)
+        result = self.cursor.fetchone()
+        if result:
+            truncate_query = f"TRUNCATE TABLE {table_name};"
+            self.cursor.execute(truncate_query)
+            self.conn.commit()
+        else:
+            return
 
-    def insert_table(self,choose):
-        table_name='func'
-        self.table_clear(table_name)
-        if choose==1:
-            f=self.config_ini['main_project']['project_path']+self.config_ini['db']['danger_funcs']
-        if choose==2:
-            f = self.config_ini['main_project']['project_path'] + self.config_ini['db']['back_up']
-        with open(f, 'r') as encrypted_file:
+    def insert_user(self,table):
+        f=self.config_ini['main_project']['project_path']+self.config_ini['db']['user']
+        with open(f,'r')as user_file:
+            if not bool(user_file.read()):
+                for line in user_file:
+                    user_info=line.split('\t')
+                    data=(user_info[0],user_info[1])
+                    sql = f"INSERT INTO {table} (id,pass_word) VALUES (%s,%s)"
+                    self.cursor.execute(sql,data)
+                    self.conn.commit()
+
+
+    def insert_func(self,table,file):
+        with open(file, 'r') as encrypted_file:
             for line in encrypted_file:
                 # 解码加密后的结果
                 encrypted_data = line.strip().split('\t')
@@ -98,8 +108,22 @@ class DB():
                 decrypted_solution = self.decrypt(solution)
 
                 data = (decrypted_name, decrypted_level, decrypted_solution)
-                sql = f"INSERT INTO {table_name} (name,level,description) VALUES (%s,%s,%s)"
+                sql = f"INSERT INTO {table} (name,level,description) VALUES (%s,%s,%s)"
                 self.cursor.execute(sql, data)
                 self.conn.commit()
+
+    def insert_table(self,choose):
+        table_name_1=self.config_ini['db_set']['form_1']
+        table_name_2=self.config_ini['db_set']['form_2']
+        self.table_clear(table_name_2)
+        if choose==1:
+            f=self.config_ini['main_project']['project_path']+self.config_ini['db']['danger_funcs']
+            self.insert_func(table_name_2,f)
+        if choose==2:
+            f = self.config_ini['main_project']['project_path'] + self.config_ini['db']['back_up']
+            self.insert_func(table_name_2,f)
+        if choose==3:
+            self.insert_user(table_name_1)
+
